completed 01-css-methods
have not checked to see if it matches solution. after pushing this commit, I will compare my attempt with the solution, and will update the readme.md with my notes, before pushing it as a commit. based on comparison will either mean i move to the next assignment or continue refining this one. however so far, it looks as if I solved it.